requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
enum34==1.1.6
fudge==1.1.0
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
pprintpp==0.3.0
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pydub==0.16.6
Pygments==2.1.3
pyOSC==0.3.5b5294
python-osc==1.6.2
requests==2.12.1
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
soundcloud==0.5.0
traitlets==4.3.1
wcwidth==0.1.7
